Restructured queue and tasks to internal folder
