+++ cmd/worker/main.go
@@ -3,8 +3,8 @@ package main
 import (
 	"context"
 	"flag"
-	"goDemoApi/cmd/worker/tasks"
-	"goDemoApi/queue"
+	"goDemoApi/internal/worker/queue"
+	"goDemoApi/internal/worker/tasks"
 	"log"
 
 	"github.com/joho/godotenv"
